Keep chimera/search query abundance 64-bit
searchinfo_s.qsize and chimera_info_s.query_size held the query
abundance as int, narrowing a ;size= above INT_MAX. This fed the abskew
/ size-ratio filtering (searchcore.cc maxsizeratio/minsizeratio and the
chimera abundance accounting, whose accumulators are already int64_t),
so once abundances were stored at full width the truncated query size
produced wrong skew decisions for very large abundances. Widen both
fields to int64_t, matching db_getabundance() and the int64_t
accumulators.

+++ src/chimera.cc
@@ -140,7 +140,7 @@ struct chimera_info_s
   int query_no = 0;
   std::vector<char> query_head;
   int query_head_len = 0;
-  int query_size = 0;
+  int64_t query_size = 0;
   std::vector<char> query_seq;
   int query_len = 0;
